Split the file
